Move route handlers to submodules.
