+++ lib/client-manager.js
@@ -7,7 +7,6 @@ var configManager = require('./config-manager');
 var backend = require('./backend');
 
 var settings = configManager.getSettings();
-
 var clientManager = {
   io: null,
   authenticatedClients: {},
@@ -19,6 +18,11 @@ var clientManager = {
   channels: {}
 };
 
+/**
+ * Initializes socket.io, adds event listeners for socket events.
+ * @param io
+ *   Socket.io as returned by require('socket.io').
+ */
 clientManager.initSocketIo = function (io) {
   this.io = io;
 
@@ -46,8 +50,9 @@ clientManager.initSocketIo = function (io) {
 };
 
 /**
- * Registers a socket.
+ * Registers a socket in the internal store.
  * @param socket
+ *   A socket object.
  */
 clientManager.addSocket = function (socket) {
   process.emit('client-connection', socket.id);
@@ -96,6 +101,13 @@ clientManager.authenticateClient = function (client, message) {
   }
 };
 
+/**
+ * Processes an incoming message event on the socket.
+ * @param socketId
+ *   The socket id that received the message.
+ * @param message
+ *   Arbitrary message object.
+ */
 clientManager.processMessage = function (socketId, message) {
   // If the message is from an active client, then process it.
   if (this.sockets[socketId] && message.hasOwnProperty('type')) {
@@ -131,6 +143,7 @@ clientManager.processMessage = function (socketId, message) {
 
 /**
  * Handle authentication call response.
+ * Note: This is a callback on an http request, so .this will not refer to this object.
  */
 clientManager.authenticateClientCallback = function (error, response, body) {
   if (error) {
@@ -323,6 +336,379 @@ clientManager.setUserOffline = function (uid) {
 };
 
 /**
+ * Kicks a user.
+ */
+clientManager.kickUser = function (uid) {
+  // Delete the user from the authenticatedClients hash.
+  for (var authToken in this.authenticatedClients) {
+    if (this.authenticatedClients[authToken].uid == uid) {
+      delete this.authenticatedClients[authToken];
+    }
+  }
+
+  // Destroy any socket connections associated with this uid.
+  for (var clientId in this.sockets) {
+    if (this.sockets[clientId].uid == uid) {
+      delete this.sockets[clientId];
+
+      if (settings.debug) {
+        console.log('kickUser: deleted socket "' + clientId + '" for uid "' + uid + '"');
+      }
+      // Delete any channel entries for this clientId.
+      for (var channel in this.channels) {
+        delete this.channels[channel].sessionIds[clientId];
+      }
+    }
+  }
+};
+
+/**
+ * Add a user to a channel.
+ * @return
+ *   True on success, false if the client is not found.
+ */
+clientManager.addUserToChannel = function (channel, uid) {
+  this.channels[channel] = this.channels[channel] || {'sessionIds': {}};
+
+  var sessionIds = this.getNodejsSessionIdsFromUid(uid);
+  if (sessionIds.length > 0) {
+    for (var i in sessionIds) {
+      this.channels[channel].sessionIds[sessionIds[i]] = sessionIds[i];
+    }
+    if (settings.debug) {
+      console.log("Added channel '" + channel + "' to sessionIds " + sessionIds.join());
+    }
+  }
+  else {
+    console.log("No active sessions for uid: " + uid);
+    return false;
+  }
+
+  // @TODO: Does this need to run even when the client is not connected?
+  // (authenticatedClients data will be reused when the connect.)
+  for (var authToken in this.authenticatedClients) {
+    if (this.authenticatedClients[authToken].uid == uid) {
+      if (this.authenticatedClients[authToken].channels.indexOf(channel) == -1) {
+        this.authenticatedClients[authToken].channels.push(channel);
+        if (settings.debug) {
+          console.log("Added channel '" + channel + "' authenticatedClients");
+        }
+      }
+    }
+  }
+
+  return true;
+};
+
+/**
+ * Get the list of Node.js sessionIds for a given authToken.
+ */
+clientManager.getNodejsSessionIdsFromAuthToken = function (authToken) {
+  var sessionIds = [];
+  for (var sessionId in this.sockets) {
+    if (this.sockets[sessionId].authToken == authToken) {
+      sessionIds.push(sessionId);
+    }
+  }
+  if (settings.debug) {
+    console.log('getNodejsSessionIdsFromAuthToken', {authToken: authToken, sessionIds: sessionIds});
+  }
+  return sessionIds;
+};
+
+/**
+ * Add an authToken to a channel.
+ * @TODO Unused, needs testing.
+ */
+clientManager.addAuthTokenToChannel = function (channel, authToken) {
+  if (!this.authenticatedClients[authToken]) {
+    console.log("Unknown authToken : " + authToken);
+    return false;
+  }
+
+  this.channels[channel] = this.channels[channel] || {'sessionIds': {}};
+  var sessionIds = this.getNodejsSessionIdsFromAuthToken(authToken);
+  if (sessionIds.length > 0) {
+    for (var i in sessionIds) {
+      this.channels[channel].sessionIds[sessionIds[i]] = sessionIds[i];
+    }
+    if (settings.debug) {
+      console.log("Added sessionIds '" + sessionIds.join() + "' to channel '" + channel + "'");
+    }
+  }
+  else {
+    console.log("No active sessions for authToken: " + authToken);
+    return false;
+  }
+
+  if (this.authenticatedClients[authToken].channels.indexOf(channel) == -1) {
+    this.authenticatedClients[authToken].channels.push(channel);
+    if (settings.debug) {
+      console.log("Added channel '" + channel + "' to authenticatedClients");
+    }
+  }
+
+  return true;
+};
+
+/**
+ * Remove an authToken from a channel.
+ * @TODO Unused, needs testing.
+ */
+clientManager.removeAuthTokenFromChannel = function (channel, authToken) {
+  if (!this.authenticatedClients[authToken]) {
+    console.log('Invalid authToken: ' + uid);
+    return false;
+  }
+  if (this.channels[channel]) {
+    var sessionIds = this.getNodejsSessionIdsFromAuthToken(authToken);
+    for (var i in sessionIds) {
+      if (this.channels[channel].sessionIds[sessionIds[i]]) {
+        delete this.channels[channel].sessionIds[sessionIds[i]];
+      }
+    }
+
+    if (this.authenticatedClients[authToken]) {
+      var index = this.authenticatedClients[authToken].channels.indexOf(channel);
+      if (index != -1) {
+        delete this.authenticatedClients[authToken].channels[index];
+      }
+    }
+
+    if (settings.debug) {
+      console.log("Successfully removed authToken '" + authToken + "' from channel '" + channel + "'.");
+    }
+    return true;
+  }
+  else {
+    console.log("Non-existent channel name '" + channel + "'");
+    return false;
+  }
+};
+
+/**
+ * Add a client (specified by session ID) to a channel.
+ * @TODO: Unused. Shall we keep it?
+ */
+clientManager.addClientToChannel = function (sessionId, channel) {
+  if (sessionId && channel) {
+    if (!/^[0-9a-z_-]+$/i.test(sessionId) || !this.sockets.hasOwnProperty(sessionId)) {
+      console.log("addClientToChannel: Invalid sessionId: " + sessionId);
+    }
+    else if (!/^[a-z0-9_]+$/i.test(channel)) {
+      console.log("addClientToChannel: Invalid channel: " + channel);
+    }
+    else {
+      this.channels[channel] = this.channels[channel] || {'sessionIds': {}};
+      this.channels[channel].sessionIds[sessionId] = sessionId;
+      if (settings.debug) {
+        console.log("Added channel '" + channel + "' to sessionId " + sessionId);
+      }
+      return true;
+    }
+  }
+  else {
+    console.log("addClientToChannel: Missing sessionId or channel name");
+  }
+  return false;
+};
+
+/**
+ * Remove a client (specified by session ID) from a channel.
+ * @TODO: Unused. Shall we keep it?
+ */
+clientManager.removeClientFromChannel = function (sessionId, channel) {
+  if (sessionId && channel) {
+    if (!/^[0-9a-z_-]+$/i.test(sessionId) || !this.sockets.hasOwnProperty(sessionId)) {
+      console.log("removeClientFromChannel: Invalid sessionId: " + sessionId);
+    }
+    else if (!/^[a-z0-9_]+$/i.test(channel) || !this.channels.hasOwnProperty(channel)) {
+      console.log("removeClientFromChannel: Invalid channel: " + channel);
+    }
+    else if (this.channels[channel].sessionIds[sessionId]) {
+      delete this.channels[channel].sessionIds[sessionId];
+      if (settings.debug) {
+        console.log("Removed sessionId '" + sessionId + "' from channel '" + channel + "'");
+      }
+      return true;
+    }
+  }
+  else {
+    console.log("removeClientFromChannel: Missing sessionId or channel name");
+  }
+  return false;
+};
+
+/**
+ * Remove a user from a channel.
+ * @return
+ *   True on success, false if the client is not found.
+ */
+clientManager.removeUserFromChannel = function (channel, uid) {
+  if (this.channels[channel]) {
+    var sessionIds = this.getNodejsSessionIdsFromUid(uid);
+    for (var i in sessionIds) {
+      if (this.channels[channel].sessionIds[sessionIds[i]]) {
+        delete this.channels[channel].sessionIds[sessionIds[i]];
+      }
+    }
+
+    for (var authToken in this.authenticatedClients) {
+      if (this.authenticatedClients[authToken].uid == uid) {
+        var index = this.authenticatedClients[authToken].channels.indexOf(channel);
+        if (index != -1) {
+          delete this.authenticatedClients[authToken].channels[index];
+        }
+      }
+    }
+
+    if (settings.debug) {
+      console.log("Successfully removed uid '" + uid + "' from channel '" + channel + "'");
+    }
+
+    return true;
+  }
+  else {
+    console.log("Non-existent channel name '" + channel + "'");
+    return false;
+  }
+};
+
+/**
+ * Add a channel.
+ */
+clientManager.addChannel = function (channel) {
+  if (this.channels[channel]) {
+    console.log("Channel name '" + channel + "' already exists.");
+    return false;
+  }
+  this.channels[channel] = {'sessionIds': {}};
+  if (settings.debug) {
+    console.log("Successfully added channel '" + channel + "'");
+  }
+};
+
+/**
+ * Returns data for healthcheck purposes.
+ */
+clientManager.getStats = function () {
+  return {
+    'authenticatedClients': Object.keys(this.authenticatedClients).length,
+    'sockets': Object.keys(this.sockets).length,
+    'onlineUsers': Object.keys(this.onlineUsers).length,
+    'tokenChannels': Object.keys(this.tokenChannels).length
+  }
+};
+
+/**
+ * Checks whether a channel exists.
+ * @param channel
+ *   Channel name.
+ * @return
+ *   True on success false on error.
+ */
+clientManager.checkChannel = function (channel) {
+  if (this.channels[channel]) {
+    console.log("Channel name '" + channel + "' is active on the server.");
+    return true;
+  }
+  else {
+    console.log("Channel name '" + channel + "' is not active on the server.");
+    return false;
+  }
+};
+
+/**
+ * Remove a channel.
+ * @param channel
+ *   Channel name.
+ * @return
+ *   True on success false on error.
+ */
+clientManager.removeChannel = function (channel) {
+  if (this.channels[channel]) {
+    delete this.channels[channel];
+    if (settings.debug) {
+      console.log("Successfully removed channel '" + channel + "'");
+    }
+    return true;
+  }
+  else {
+    console.log("Non-existent channel name '" + channel + "'");
+    return false;
+  }
+};
+
+/**
+ * Set the list of users a uid can see presence info about.
+ */
+clientManager.setUserPresenceList = function (uid, uidlist) {
+  for (var i in uidlist) {
+    if (!/^\d+$/.test(uidlist[i])) {
+      console.log("Invalid uid: " + uid);
+      return false;
+    }
+  }
+  this.onlineUsers[uid] = uidlist;
+};
+
+/**
+ * Get the list of backend uids and authTokens connected to a content token channel.
+ */
+clientManager.getContentTokenChannelUsers = function (channel) {
+  var users = {uids: [], authTokens: []};
+  for (var sessionId in this.tokenChannels[channel].sockets) {
+    if (this.sockets[sessionId].uid) {
+      users.uids.push(this.sockets[sessionId].uid);
+    }
+    else {
+      users.authTokens.push(this.sockets[sessionId].authToken);
+    }
+  }
+  return users;
+};
+
+/**
+ * Set a content token.
+ */
+clientManager.setContentToken = function (channel, token, value) {
+  this.tokenChannels[channel] = this.tokenChannels[channel] || {'tokens': {}, 'sockets': {}};
+  this.tokenChannels[channel].tokens[token] = value;
+};
+
+/**
+ * Publish a message to clients subscribed to a channel.
+ * @return
+ *   True on success; false on error.
+ */
+clientManager.publishMessageToContentChannel = function (channel, message) {
+  if (!this.tokenChannels.hasOwnProperty(channel)) {
+    console.log('publishMessageToContentChannel: The channel "' + channel + '" doesn\'t exist.');
+    return false;
+  }
+
+  for (var socketId in this.tokenChannels[channel].sockets) {
+    this.publishMessageToClient(socketId, message);
+  }
+  return true;
+};
+
+/**
+ * Logout the given user from the server.
+ */
+clientManager.logoutUser = function (authToken) {
+  // Delete the user from the authenticatedClients hash.
+  delete this.authenticatedClients[authToken];
+
+  // Destroy any socket connections associated with this authToken.
+  for (var clientId in this.sockets) {
+    if (this.sockets[clientId].authToken == authToken) {
+      this.cleanupSocket(this.sockets[clientId]);
+    }
+  }
+};
+
+/**
  * Check for any open sockets associated with the channel and socket pair.
  */
 clientManager.checkTokenChannelStatus = function (tokenChannel, socket) {
@@ -398,10 +784,16 @@ clientManager.publishMessageToChannel = function (message) {
   return clientCount;
 };
 
+/**
+ * Broadcasts a message to all sockets.
+ */
 clientManager.broadcastMessage = function (message) {
   this.io.sockets.json.send(message);
 };
 
+/**
+ * Returns the number of open sockets.
+ */
 clientManager.getSocketCount = function () {
   return this.io.sockets.sockets.length;
 };
